Urine-Test/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
blinker==1.8.2
boto3==1.35.49
botocore==1.35.49
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
distro==1.9.0
Flask==3.0.3
fpdf==1.7.2
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.6.1
jmespath==1.0.1
MarkupSafe==3.0.2
multidict==6.1.0
openai==0.27.0
propcache==0.2.0
pydantic==2.9.2
pydantic_core==2.23.4
PyMuPDF==1.24.12
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
s3transfer==0.10.3
six==1.16.0
sniffio==1.3.1
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.6
yarl==1.16.0
